requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow
gdown
scikit-image
python-multipart

# fastapi
# uvicorn[standard]
# tensorflow
# pillow
# gdown
# scikit-image
# python-multipart
